requirements.txt
streamlit
langchain==0.3.27
langchain-openai
langchain-community
langchain-huggingface
sentence-transformers
chromadb
pypdf
python-docx
tiktoken
# pysqlite3-binary
rapidocr-onnxruntime
rapidocr-pdf
scikit-learn
langchain-ollama
langchain-openai
langchain-google-genai
